Zendesk_Scripts/Ticket_work/bulk_ticket_export/automated_analysis/requirements.txt
# Python dependencies for automated ticket analysis

# Google Gemini API
google-generativeai>=0.8.0

# HTTP requests (used by zendesk_exporter.py)
requests>=2.31.0
